requirements/local.txt
# Local Dependencies
-r base.txt
django-finalware